% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamnetwork-package.R
\docType{package}
\name{streamnetwork-package}
\alias{streamnetwork}
\alias{streamnetwork-package}
\title{streamnetwork: Create, Manipulate, Query and Plot Stream Networks}
\description{
An R package to create, manipulate, query and plot stream networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/streamnetwork}
  \item Report bugs at \url{https://github.com/poissonconsulting/streamnetwork/issues}
}

}
\author{
\strong{Maintainer}: Poisson Consulting \email{admin@poissonconsulting.ca} [copyright holder]

}
\keyword{internal}
